CE-2783: Added Template Classification Service
